Bug 1650299 - Unify the inclusion of the ICU data file. r?froydnj
All the supported compilers support a GNU AS-like syntax, with only a
few details varying. It means we can use a single, simpler, way to
include the ICU data file, instead of 3 different ways, including one
that uses armasm64.exe, possibly wrapped with Wine.

Differential Revision: https://phabricator.services.mozilla.com/D82144

Gbp-Pq: Topic fixes
Gbp-Pq: Name Bug-1650299-Unify-the-inclusion-of-the-ICU-data-file.patch